“TopicHeaderImageをS3へ直接アップロード”

+++ app/controllers/topics_controller.rb
@@ -1,9 +1,10 @@
 class TopicsController < ApplicationController
   skip_before_action :authenticate_user!, only: %i[index show]
+  before_action :set_s3_direct_post, only: [:new, :edit, :create, :update]
 
   def index
-    @search = Topic.includes([header_image_attachment: :blob], { user: [image_attachment: :blob] }, :netas).where(private_flag: false).ransack(params[:q])
-    @topic_cards = @search.result(distinct: true).order('created_at DESC').page(params[:page]).per(5)
+    @search = Topic.includes({ user: [image_attachment: :blob] }, :netas).where(private_flag: false).ransack(params[:q])
+    @topic_cards = @search.result(distinct: true).order('created_at DESC').page(params[:page]).per(20)
     @hashtag_ranking = Hashtag.get_ranking(10)
   end
 
@@ -43,6 +44,7 @@ class TopicsController < ApplicationController
   def update
     @topic = Topic.find(params[:id])
     if owner(@topic)
+      delete_s3_object
       if @topic.update(post_params)
         redirect_to topic_path(@topic.id), notice: 'トピックを更新しました。' and return
       else
@@ -56,10 +58,14 @@ class TopicsController < ApplicationController
   def destroy
     @topic = Topic.includes(:netas, :pageviews, :bookmarks).find(params[:id])
     if owner(@topic)
-      if @topic.destroy
-        redirect_to topics_path, notice: 'トピックを削除しました。' and return
+      if delete_s3_object
+        if @topic.destroy
+          redirect_to topics_path, notice: 'トピックを削除しました。' and return
+        else
+          render :edit and return
+        end
       else
-        render :edit and return
+        redirect_to topic_path(@topic.id), alert: 'ヘッダー画像が削除できません。' and return
       end
     else
       redirect_to topic_path(@topic.id), alert: '権限がありません。' and return
@@ -69,7 +75,7 @@ class TopicsController < ApplicationController
   private
 
   def post_params
-    params.require(:topic).permit(:title, :content, :header_image, :private_flag).merge(user_id: current_user.id)
+    params.require(:topic).permit(:title, :content, :header_img_url, :private_flag).merge(user_id: current_user.id)
   end
 
   def owner(topic)
@@ -77,6 +83,27 @@ class TopicsController < ApplicationController
       topic.owner(current_user)
     else
       false
+    end
+  end
+
+  def set_s3_direct_post
+    @s3_direct_post = S3_BUCKET.presigned_post(key: "topic_header_images/#{SecureRandom.uuid}/${filename}", success_action_status: '201', acl: 'public-read')
+  end
+
+  def delete_s3_object
+    if @topic.header_img_url.present?
+      if @topic.header_img_url.include?('amazonaws.com/')
+        object = S3_BUCKET.object(@topic.header_img_url.split('amazonaws.com/')[1])
+        if object.present?
+          object.delete
+        else
+          false
+        end
+      else
+        false
+      end
+    else
+      true
     end
   end
 end
